Updated alphabet plus on night_writer.rb
