+++ message.txt
@@ -1,5 +1,1 @@
-hello world
-She was more like a beauty queen from a movie scene
-I said don't mind, but what do you mean, I am the one
-Who will dance on the floor in the round?
-She said I am the one, who will dance on the floor in the round
+ab
