+++ spec/night_writer_spec.rb
@@ -1,0 +1,10 @@
+require './lib/night_writer'
+
+RSpec.describe NightWriter do
+  it 'exists' do
+    night_writer = NightWriter.new
+    expect(night_writer).to be_a(NightWriter)
+  end
+
+
+end
